+++ cmd/run.go
@@ -112,3 +112,25 @@ func RunDiscard(ctx *cli.Context) (int, error) {
 
 	return 0, nil
 }
+
+// RunCurrentID return the ID of the current run on TFC
+func RunCurrentID(ctx *cli.Context) (int, error) {
+	c, cfg, err := configure(ctx)
+	if err != nil {
+		return 1, err
+	}
+
+	w, err := c.GetWorkspace(cfg.Runtime.TFC.Organization, cfg.Runtime.TFC.Workspace)
+	if err != nil {
+		return 1, err
+	}
+
+	runID, err := c.GetWorkspaceCurrentRunID(w)
+	if err != nil {
+		return 1, err
+	}
+
+	fmt.Println(runID)
+
+	return 0, nil
+}